+++ Application/Users/UseCases/ListUsers/ListUsersHandler.cs
@@ -44,7 +44,7 @@ public sealed class ListUsersHandler(
             FullName: user.FullName.Value,
             Email: user.Email.Value,
             BirthDate: user.BirthDate.Value,
-            Role: user.Role,
+            Role: (int)user.Role,
             MustChangePassword: user.MustChangePassword,
             CreatedAt: user.CreatedAt,
             UpdatedAt: user.UpdatedAt)).ToList();
